Playground/backend/requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-fireworks>=0.1.0
voyageai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pymongo>=4.6.0
langsmith>=0.1.0

# RAG System Dependencies
chromadb>=0.4.22
pymupdf>=1.23.0

